Added intial files for the texture task

+++ OpenGL-Practice/HelloShaders.h
@@ -113,7 +113,7 @@ public:
 	};
 	void setFloat(const std::string& name, float value) const
 	{
-		glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+		glUniform1f(glGetUniformLocation(ID, name.c_str()), (float)value);
 	};
 };
